custom-elements/registering-custom-elements/extensions-to-document-interface/custom-element-name.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element local name is the value of the EXTENDS property</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Custom element local name is the value of the EXTENDS property">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-register">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function () {
    var doc = newHTMLDocument();
    
    HTML5_ELEMENTS.forEach(function(element) {
        var name = 'x-' + element;
        var obj = doc.createElement(element);
        var proto = Object.create(obj.constructor.prototype);
        var GeneratedConstructor = doc.register(name, {
            prototype: proto,
            extends: element
        });
        var customElement = new GeneratedConstructor();

        assert_equals(customElement.localName, element,
                'Custom element local name must be an EXTENDS value');        
    });
}, 'Custom element local name is the value of the EXTENDS property');

</script>
</body>
</html>
